+++ Koha/Plugin/Com/ByWaterSolutions/RapidoILL/lib/RapidoILL/AgencyPatron.pm
@@ -1,0 +1,13 @@
+package RapidoILL::AgencyPatron;
+
+use Modern::Perl;
+
+no warnings 'redefine';
+
+use base qw(Koha::Object);
+
+sub _type {
+    return 'KohaPluginComBywatersolutionsRapidoillAgencyToPatron';
+}
+
+1;